+++ packages/labs/testing/src/index.ts
@@ -5,4 +5,5 @@
  */
 
 export {litSsrPlugin} from './lib/litSsrPlugin.js';
-export * from './lib/fixtures.js';
+export * from './lib/fixtures/csrFixture.js';
+export * from './lib/fixtures/ssrFixture.js';
