+++ packages/labs/testing/src/lib/fixtures/ssrFixture.ts
@@ -10,7 +10,7 @@ import {hydrateShadowRoots} from '@webcomponents/template-shadowroot';
 import {litSsrPluginCommand} from '../constants.js';
 
 import type {LitElement, TemplateResult} from 'lit';
-import type {SsrFixtureOption} from './fixtureOption.js';
+import type {FixtureOption, SsrFixtureOption} from './fixtureOption.js';
 import type {Payload} from '../litSsrPlugin.js';
 
 // Enhance DOMParser's parseFromString method to include `includeShadowRoots`
@@ -77,4 +77,47 @@ export async function ssrFixture(
     await el.updateComplete;
   }
   return el;
+}
+
+/**
+ * Renders the provided lit-html template with a Lit element server-side by
+ * executing a custom command for Web Test Runner provided by the Lit SSR
+ * Plugin, loads it to the document and hydrates it, returning the element.
+ *
+ * This module **must** be imported before any custom element definitions.
+ *
+ * @param {TemplateResult} template - lit-html template. Must contain a single
+ * top level custom element.
+ * @param {string[]} option.modules - Path to custom element definition modules
+ * needed to render template, relative to the project root.
+ * @param {string} option.base - Base path for the module. Generally should be
+ * `import.meta.url`.
+ */
+export async function ssrHydratedFixture(
+  template: TemplateResult,
+  {modules, base}: FixtureOption
+) {
+  return ssrFixture(template, {modules, base, hydrate: true});
+}
+
+/**
+ * Renders the provided lit-html template with a Lit element server-side by
+ * executing a custom command for Web Test Runner provided by the Lit SSR
+ * Plugin, loads it to the document **without** hydrating it, returning the
+ * element.
+ *
+ * This module **must** be imported before any custom element definitions.
+ *
+ * @param {TemplateResult} template - lit-html template. Must contain a single
+ * top level custom element.
+ * @param {string[]} option.modules - Path to custom element definition modules
+ * needed to render template, relative to the project root.
+ * @param {string} option.base - Base path for the module. Generally should be
+ * `import.meta.url`.
+ */
+export async function ssrNonHydratedFixture(
+  template: TemplateResult,
+  {modules, base}: FixtureOption
+) {
+  return ssrFixture(template, {modules, base, hydrate: false});
 }
